impression_backend/requirements.txt
Django
django-cors-headers
psycopg2-binary
mysqlclient
whitenoise
djangorestframework
django-role-permissions
